Create morphology subpackage
