+++ photutils/morphology/__init__.py
@@ -1,0 +1,7 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+"""
+This subpackage contains tools for measuring morphological properties of
+objects in an astronomical image.
+"""
+
+from .core import *
